fix: deleted types.ts and moved rootprops to index.ts
